# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestión de compañías y empleados",
        "title": "Compania API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/companias": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compañías"
                ],
                "summary": "Obtener todas las compañías",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compañías"
                ],
                "summary": "Crear una compañía",
                "parameters": [
                    {
                        "description": "Datos de la compañía",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateCompaniaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/companias/con-empleados": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compañías"
                ],
                "summary": "Crear compañía con empleados",
                "parameters": [
                    {
                        "description": "Compañía con empleados",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateCompaniaWithEmpleadosRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/companias/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compañías"
                ],
                "summary": "Obtener una compañía por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la compañía",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compañías"
                ],
                "summary": "Actualizar una compañía",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la compañía",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la compañía",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateCompaniaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compañías"
                ],
                "summary": "Eliminar una compañía",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la compañía",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/companias/{id}/empleados": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compañías"
                ],
                "summary": "Obtener empleados de una compañía",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la compañía",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.CreateCompaniaEmpleadoRequest": {
            "type": "object",
            "required": [
                "apellido",
                "cargo",
                "correo",
                "nombre",
                "salario"
            ],
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "cargo": {
                    "type": "string"
                },
                "correo": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "salario": {
                    "type": "number"
                }
            }
        },
        "dtos.CreateCompaniaRequest": {
            "type": "object",
            "required": [
                "direccion",
                "nombre",
                "telefono"
            ],
            "properties": {
                "direccion": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "telefono": {
                    "type": "string"
                }
            }
        },
        "dtos.CreateCompaniaWithEmpleadosRequest": {
            "type": "object",
            "required": [
                "direccion",
                "empleados",
                "nombre",
                "telefono"
            ],
            "properties": {
                "direccion": {
                    "type": "string"
                },
                "empleados": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateCompaniaEmpleadoRequest"
                    }
                },
                "nombre": {
                    "type": "string"
                },
                "telefono": {
                    "type": "string"
                }
            }
        },
...
        "dtos.UpdateCompaniaRequest": {
            "type": "object",
            "required": [
                "direccion",
                "nombre",
                "telefono"
            ],
            "properties": {
                "direccion": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "telefono": {
                    "type": "string"
                }
            }
        },
...
    }
}